Manapool visuals update

+++ Assets/Scripts/Hero/AttributesManager.cs
@@ -27,15 +27,15 @@ public class AttributesManager : MonoBehaviour
     //Attributes
     public float fullHealth = 0;
     public float hp = 0;
-    public int fullMana = 0;
-    public int mp = 0;
+    public float fullMana = 0;
+    public float mp = 0;
     public float strength = 0;
     public float luck = 0;
     public float satiety = 0;
 
     //Max values for attributes
     private readonly float maxHealth = 100000;
-    private readonly int maxMana = 100;
+    private readonly float maxMana = 100;
     private readonly float maxStrengtht = 200f;
     private readonly float maxLuck = 100f;
     private readonly float maxSatiety = 100f;
@@ -43,7 +43,7 @@ public class AttributesManager : MonoBehaviour
     //Min values for attributes
     private readonly float minMaxHealth = 1;
     private readonly float minHp = 0;
-    private readonly int minMana = 0;
+    private readonly float minMana = 0;
     private readonly float minStrength = 0;
     private readonly float minLuck = 0;
     private readonly float minSatiety = 0;
@@ -51,8 +51,8 @@ public class AttributesManager : MonoBehaviour
     //Stored value
     public float s_Health;
     public float s_hp;
-    public int s_Mana;
-    public int s_mp;
+    public float s_Mana;
+    public float s_mp;
     public float s_Strength;
     public float s_Luck;
     public float s_Satiety;
